app/backend/requirements.txt
Django~=4.1.0
djangorestframework~=3.14.0
django-cors-headers~=3.13.0
dj-database-url~=1.0.0
psycopg2-binary~=2.9.0
djangorestframework-simplejwt~=5.2.0
